--- src/modules/dashboard/components/Dialogs/EditExpenditureDialog/EditExpenditureDialog.css
@@ -1,231 +1,0 @@
-@value submitButtonWidth: 160px;
-
-.title {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 9px 0 20px;
-  width: 100%;
-  font-size: var(--size-medium-l);
-  font-weight: var(--weight-bold);
-  color: var(--dark);
-}
-
-.forceContainer {
-  display: flex;
-  align-items: center;
-  font-size: var(--size-smallish);
-}
-
-.stakeContainer {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 13px 0;
-  border-bottom: 1px solid var(--grey-blue-0);
-  font-size: var(--size-smallish);
-  font-weight: var(--weight-bold);
-  color: var(--temp-grey-blue-7);
-  opacity: 0.85;
-}
-
-.messageContainer {
-  padding-bottom: 90px;
-  color: var(--temp-grey-blue-7);
-}
-
-.label {
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-  padding: 0;
-  font-size: var(--size-smallish);
-  font-weight: var(--weight-bold);
-  color: color-mod(var(--temp-grey-blue-7) alpha(85%));
-  letter-spacing: 0.5px;
-}
-
-.icon figure {
-  display: block;
-}
-
-.tokenIcon {
-  margin-right: 6px;
-  height: 16px;
-  width: 16px;
-}
-
-.symbol {
-  margin-left: 4px;
-}
-
-.row {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 8px 0 13px;
-  font-size: var(--size-smallish);
-  font-weight: var(--weight-bold);
-  line-height: 18px;
-  color: color-mod(var(--temp-grey-blue-7) alpha(85%));
-}
-
-.valueLabel {
-  align-items: flex-start;
-}
-
-.smallerPadding {
-  padding: 4px 0 8px;
-}
-
-.withoutPadding {
-  padding: 0;
-}
-
-.forceRow {
-  align-items: baseline;
-  padding-top: 16px;
-  white-space: nowrap;
-}
-
-.recipientContainer {
-  display: flex;
-  justify-content: left;
-  align-items: center;
-  font-size: var(--size-smallish);
-  font-weight: var(--weight-bold);
-  line-height: 18px;
-  color: color-mod(var(--dark) alpha(85%));
-}
-
-.recipientContainer span {
-  margin-left: 4px;
-  color: inherit;
-}
-
-.recipientContainer div {
-  color: inherit;
-}
-
-.reicpientButtonContainer button {
-  padding: 0;
-  height: auto;
-  color: var(--action-secondary);
-}
-
-.reicpientButtonContainer {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 27px 0 8px;
-}
-
-.valueContainer {
-  text-align: right;
-}
-
-.value {
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-  font-size: var(--size-normal);
-  font-weight: var(--weight-normal);
-  color: var(--dark);
-}
-
-.value > span {
-  margin-right: 4px;
-}
-
-.paddingBottom {
-  padding-bottom: 6px;
-}
-
-.userAvatarContainer {
-  display: flex;
-  align-items: center;
-}
-
-.userAvatarContainer > span {
-  margin-left: 8px;
-}
-
-.contentWrapper {
-  padding: 0 20px 54px;
-}
-
-.descriptionWrapper {
-  padding-top: 10px;
-}
-
-.valueAmount {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  color: var(--dark);
-}
-
-.changeContainer {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 27px 0 13px;
-  font-size: var(--size-smallish);
-  font-weight: var(--weight-bold);
-  line-height: 18px;
-  color: var(--dark);
-}
-
-.changeItem {
-  padding: 8px 0 13px;
-  color: var(--temp-grey-blue-7);
-}
-
-.changeContainer button {
-  padding: 0;
-  height: auto;
-  color: var(--action-secondary);
-}
-
-.changeWrapper {
-  max-width: 70%;
-  overflow: elipsis;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-
-.teamWrapper {
-  display: flex;
-  align-items: center;
-}
-
-.teamWrapper > div {
-  margin-right: 4px;
-}
-
-.noChanges {
-  font-weight: var(--weight-bold);
-  color: var(--pink);
-}
-
-.annotationsWrapper {
-  margin-bottom: 11px;
-}
-
-.toggleContainer i {
-  margin: 0 -20px 4px 21px;
-
-  /* next lines are used for the correct tooltip arrow placement */
-  transform: translateX(-25px);
-}
-
-/* this selector is for tooltip wrapper */
-.toggleContainer > div > div:last-child {
-  max-width: 250px;
-  white-space: normal;
-}
-
-.toggleContainer label {
-  margin-bottom: 4px;
-}
