--- src/modules/dashboard/components/Dialogs/EditExpenditureDialog/EditExpenditureDialog.css.d.ts
@@ -1,30 +1,0 @@
-export const submitButtonWidth: string;
-export const title: string;
-export const forceContainer: string;
-export const stakeContainer: string;
-export const messageContainer: string;
-export const label: string;
-export const icon: string;
-export const tokenIcon: string;
-export const symbol: string;
-export const row: string;
-export const valueLabel: string;
-export const smallerPadding: string;
-export const withoutPadding: string;
-export const forceRow: string;
-export const recipientContainer: string;
-export const reicpientButtonContainer: string;
-export const valueContainer: string;
-export const value: string;
-export const paddingBottom: string;
-export const userAvatarContainer: string;
-export const contentWrapper: string;
-export const descriptionWrapper: string;
-export const valueAmount: string;
-export const changeContainer: string;
-export const changeItem: string;
-export const changeWrapper: string;
-export const teamWrapper: string;
-export const noChanges: string;
-export const annotationsWrapper: string;
-export const toggleContainer: string;
