+++ src/modules/dashboard/components/Dialogs/EditExpenditureDialog/EditExpenditureDialogForm.css.d.ts
@@ -1,0 +1,14 @@
+export const submitButtonWidth: string;
+export const title: string;
+export const withoutPadding: string;
+export const forceRow: string;
+export const contentWrapper: string;
+export const descriptionWrapper: string;
+export const changeContainer: string;
+export const changeItem: string;
+export const changeWrapper: string;
+export const teamWrapper: string;
+export const noChanges: string;
+export const annotationsWrapper: string;
+export const toggleContainer: string;
+export const row: string;
