+++ src/modules/dashboard/components/Dialogs/EditExpenditureDialog/NewDelay/NewDelay.css
@@ -1,0 +1,29 @@
+.row {
+  composes: row from '../EditExpenditureDialogForm.css';
+}
+
+.label {
+  display: flex;
+  justify-content: flex-end;
+  align-items: center;
+  padding: 0;
+  font-size: var(--size-smallish);
+  font-weight: var(--weight-bold);
+  color: color-mod(var(--temp-grey-blue-7) alpha(85%));
+  letter-spacing: 0.5px;
+}
+
+.value {
+  display: flex;
+  justify-content: flex-end;
+  align-items: center;
+  font-size: var(--size-normal);
+  font-weight: var(--weight-normal);
+  color: var(--dark);
+  letter-spacing: 0.1px;
+}
+
+.value > span {
+  margin-right: 4px;
+  letter-spacing: 0.1px;
+}
