+++ src/modules/dashboard/components/Dialogs/EditExpenditureDialog/NewDelay/NewDelay.css.d.ts
@@ -1,0 +1,3 @@
+export const row: string;
+export const label: string;
+export const value: string;
