+++ src/modules/dashboard/components/Dialogs/EditExpenditureDialog/NewDelay/index.ts
@@ -1,0 +1,1 @@
+export { default } from './NewDelay';
