+++ src/modules/dashboard/components/Dialogs/EditExpenditureDialog/NewRecipient/NewRecipient.css
@@ -1,0 +1,24 @@
+.row {
+  composes: row from '../EditExpenditureDialogForm.css';
+}
+
+.label {
+  composes: label from '../NewDelay/NewDelay.css';
+}
+
+.smallerPadding {
+  padding: 4px 0 8px;
+}
+
+.valueContainer {
+  text-align: right;
+}
+
+.userAvatarContainer {
+  display: flex;
+  align-items: center;
+}
+
+.userAvatarContainer > span {
+  margin-left: 8px;
+}
