+++ src/modules/dashboard/components/Dialogs/EditExpenditureDialog/NewRecipient/NewRecipient.css.d.ts
@@ -1,0 +1,5 @@
+export const row: string;
+export const label: string;
+export const smallerPadding: string;
+export const valueContainer: string;
+export const userAvatarContainer: string;
