+++ src/modules/dashboard/components/Dialogs/EditExpenditureDialog/NewRecipient/index.ts
@@ -1,0 +1,1 @@
+export { default } from './NewRecipient';
