+++ src/modules/dashboard/components/Dialogs/EditExpenditureDialog/NewValue/NewValue.css.d.ts
@@ -1,0 +1,9 @@
+export const row: string;
+export const label: string;
+export const value: string;
+export const smallerPadding: string;
+export const valueContainer: string;
+export const userAvatarContainer: string;
+export const valueLabel: string;
+export const paddingBottom: string;
+export const tokenIcon: string;
