+++ src/modules/dashboard/components/Dialogs/EditExpenditureDialog/NewValue/index.ts
@@ -1,0 +1,1 @@
+export { default } from './NewValue';
