+++ src/modules/dashboard/components/Dialogs/EditExpenditureDialog/Recipient/Recipient.css.d.ts
@@ -1,0 +1,7 @@
+export const reicpientButtonContainer: string;
+export const recipientContainer: string;
+export const valueAmount: string;
+export const tokenIcon: string;
+export const icon: string;
+export const symbol: string;
+export const row: string;
