+++ src/modules/dashboard/components/ExpenditurePage/Payments/CollapseExpandButtons.tsx
@@ -1,9 +1,21 @@
 import React from 'react';
 import classNames from 'classnames';
+import { defineMessages } from 'react-intl';
 
 import Icon from '~core/Icon';
+
 import styles from './Payments.css';
-import { MSG } from './Payments';
+
+const MSG = defineMessages({
+  minusIconTitle: {
+    id: `dashboard.ExpenditurePage.Payments.CollapseExpandButtons.minusIconTitle`,
+    defaultMessage: 'Collapse a single recipient settings',
+  },
+  plusIconTitle: {
+    id: `dashboard.ExpenditurePage.Payments.CollapseExpandButtons.plusIconTitle`,
+    defaultMessage: 'Expand a single recipient settings',
+  },
+});
 
 interface Props {
   isExpanded: boolean;
@@ -12,7 +24,7 @@ interface Props {
   isLocked?: boolean;
 }
 
-const displayName = 'dashboard.ExpenditurePage.CollapseExpandButtons';
+const displayName = 'dashboard.ExpenditurePage.Payments.CollapseExpandButtons';
 
 const CollapseExpandButtons = ({
   isExpanded,
