+++ src/modules/dashboard/components/ExpenditurePage/Payments/Payments.css
@@ -138,7 +138,3 @@
   border-radius: 50%;
   background-color: var(--golden);
 }
-
-.blue {
-  background-color: var(--action-secondary);
-}
