+++ src/modules/dashboard/components/ExpenditurePage/Payments/Payments.css.d.ts
@@ -16,4 +16,3 @@ export const recipientName: string;
 export const editIcon: string;
 export const recipientNameWrapper: string;
 export const dot: string;
-export const blue: string;
