+++ src/modules/dashboard/components/ExpenditurePage/Payments/RecipientHeader/RecipientHeader.css.d.ts
@@ -1,0 +1,4 @@
+export const recipientNameWrapper: string;
+export const recipientName: string;
+export const dot: string;
+export const blue: string;
