+++ src/modules/dashboard/components/ExpenditurePage/Payments/RecipientHeader/index.ts
@@ -1,0 +1,1 @@
+export { default } from './RecipientHeader';
