+++ src/modules/dashboard/components/ExpenditurePage/Stages/StageItem/index.ts
@@ -1,0 +1,1 @@
+export { default } from './StageItem';
