+++ src/modules/dashboard/components/ExpenditurePage/Stages/index.ts
@@ -1,4 +1,3 @@
 export { default } from './Stages';
-export { default as StageItem } from './StageItem/StageItem';
 export { default as LockedStages } from './LockedStages';
 export { default as FormStages } from './FormStages';
